# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CreateProductoRequest:
    properties:
      categoria_id:
        type: string
      codigo:
        type: string
      codigo_barras:
        type: string
      descripcion:
        type: string
      nombre:
        type: string
      precio:
        type: number
      stock_actual:
        type: integer
      stock_minimo:
        type: integer
    required:
    - codigo
    - nombre
    - precio
    type: object
...
  model.Producto:
    properties:
      categoria_id:
        type: string
      categoria_nombre:
        type: string
      codigo:
        type: string
      codigo_barras:
        type: string
      created_at:
        type: string
      descripcion:
        type: string
      id:
        type: string
      imagen:
        type: string
      nombre:
        type: string
      precio:
        type: number
      stock_actual:
        type: integer
      stock_minimo:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.UpdateProductoRequest:
    properties:
      categoria_id:
        type: string
      codigo:
        type: string
      codigo_barras:
        type: string
      descripcion:
        type: string
      nombre:
        type: string
      precio:
        type: number
      stock_minimo:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@stockapp.com
    name: API Support
    url: http://www.stockapp.com/support
  description: API para gestión de inventario y stock
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: StockApp API
  version: "1.0"
paths:
...
  /productos:
    get:
      description: Get all products with pagination and filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by name or code
        in: query
        name: search
        type: string
      - description: Filter by category
        in: query
        name: categoria_id
        type: string
      - description: Filter by low stock
        in: query
        name: stock_bajo
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get All Productos
      tags:
      - Productos
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: Producto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateProductoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Producto'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Producto
      tags:
      - Productos
  /productos/{id}:
    delete:
      description: Delete a product
      parameters:
      - description: Producto ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Producto
      tags:
      - Productos
    get:
      description: Get a product by ID
      parameters:
      - description: Producto ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Producto'
      security:
      - BearerAuth: []
      summary: Get Producto by ID
      tags:
      - Productos
    put:
      consumes:
      - application/json
      description: Update a product
      parameters:
      - description: Producto ID
        in: path
        name: id
        required: true
        type: string
      - description: Producto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProductoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Producto'
      security:
      - BearerAuth: []
      summary: Update Producto
      tags:
      - Productos
  /productos/{id}/imagen:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image for a product
      parameters:
      - description: Producto ID
        in: path
        name: id
        required: true
        type: string
      - description: Product image
        in: formData
        name: imagen
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload Product Image
      tags:
      - Productos
  /productos/barcode/{codigo}:
    get:
      description: Get a product by barcode code
      parameters:
      - description: Barcode code
        in: path
        name: codigo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Producto'
      security:
      - BearerAuth: []
      summary: Get Producto by Barcode
      tags:
      - Productos
  /productos/stock-bajo:
    get:
      description: Get all products with stock below minimum
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Low Stock Products
      tags:
      - Productos
  /reportes/dashboard-avanzado:
    get:
      description: Get advanced dashboard with detailed stats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      security:
      - BearerAuth: []
      summary: Get Advanced Dashboard
      tags:
      - Reportes
  /reportes/movimientos-fecha:
    get:
      description: Get movements in a date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: desde
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: hasta
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      security:
      - BearerAuth: []
      summary: Get Movimientos by Date Range
      tags:
      - Reportes
  /reportes/stock-categoria:
    get:
      description: Get stock grouped by category
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      security:
      - BearerAuth: []
      summary: Get Stock by Category
      tags:
      - Reportes
schemes:
- http
- https
swagger: "2.0"